Move response classes to response package
