--- src/main/java/com/etraveli/movierental/dto/MovieListResponse.java
@@ -1,6 +1,0 @@
-package com.etraveli.movierental.dto;
-
-import java.util.List;
-
-public record MovieListResponse(List<MovieDTO> movies) {
-}
