+++ src/main/java/com/etraveli/movierental/dto/response/MovieListResponse.java
@@ -1,0 +1,8 @@
+package com.etraveli.movierental.dto.response;
+
+import com.etraveli.movierental.dto.MovieDTO;
+
+import java.util.List;
+
+public record MovieListResponse(List<MovieDTO> movies) {
+}
